Added a new 'click_and_wait_until_gone' convenience method for Selenium.
Maybe a little specific, but you call it like:

  @driver.click_and_wait_until_gone(:id => "createPlusOne")

It will find the matching element, click it, then poll until that selector
returns a different object.  This is useful for forms where clicking the submit
button uses AJAX to replace the form--the form element IDs might be identical,
but their object identities are different.