Added support for webhooks, allowing the backend to push messages to the frontend.
To keep the UI snappy, we minimise the number of backend requests made for
each frontend request.  Some bits of information from the backend are used in
almost every page.  Currently:

  * The list of repositories

  * Subject term vocabularies

  * The list of permissions assigned to the current user

To avoid having to refetch these lists on every request, we want to allow the
frontend to cache them.  Then the usual problems of cache invalidation pop up:
we need some way for the backend to signal to the frontend when its data has
changed.

This is what we use webhooks for: the backend knows the URL of the frontend,
and registers it to receive notifications for certain events.  When backend
code calls Webhooks.notify, the backend sends a POST request to a controller on
the frontend with one or more events.  The frontend can then invalidate its
caches in response to this.

So, summarising the changes, on the backend we have:

  * A new endpoint that API clients can use to register their webhook URLs

  * A new database table for storing the list of webhook URLs

  * A new Webhooks class that allows you to broadcast an event (with
    very rudimentary support for deduplication of events and queueing)

  * Notifications fired in response to the backend starting up, groups being
    changed, repositories being added.

And on the frontend:

  * A new controller for receiving webhook notifications

  * A MemoryLeak class (heh) that provides shared memory for all threads.
    Allows getting and storing cached values in a thread-safe way.  Also
    provides support for time-based cache expiration.

  * New handlers for webhook notifications that invalidate the caches as
    appropriate.