Reworked the way the global repository is identified.
Previously I was giving the global repository a fixed ID of (2^31)-1
by manually inserting the row when the system started up.  However, it
seems that doing this and then relying on the value of the
autoincrement ID column to start from 1 is over-presumptuous.

I've switched the global repository to be identified by its repository
code now instead.