Update to the latest RC of 1.7.0
I hit issues with the backend:test suite not running locally.  It would
periodically fail with:

     [java]        uninitialized constant Rack::MockRequest

on pretty much every test.  After some digging I found that all of the failing
constants were defined by "autoload", which led me to:

  https://jira.codehaus.org/browse/JRUBY-6658

Since this is fixed in the latest JRuby, I'm upgrading to that version.  We
might as well put the new version through its paces now anyway.