Require all JSONmodel instances to include a "jsonmodel_type" indicating their record type.
So far our code has assumed that a function like this one:

  def self.map_hash_with_schema(hash, schema = nil, transformations = [])

is able to walk over a hash object and JSON schema in lockstep, matching up the
members of the hash with the corresponding entry in the schema that constrains
their types.

This breaks down where we have a hash entry like this:

  "entries" => [{"title" => "Pride and Prejudice"},
                {"type" => "Icecream cake"},
                {"breed" => "Dachshund"}]

and a schema like this:

  "entries" => {
    "type" => "array",
    "items" => [{"type" => "JSONModel(:book)"},
                {"type" => "JSONModel(:dessert)"},
                {"type" => "JSONModel(:dog)"}]
  }

Now we can't deterministically walk the tree, because we don't know which array
element corresponds to which possible type.  For validation the ambiguity
doesn't matter: if the hash validates against any of the types, then it's
valid.  But when we're doing more complicated things like applying default
values from the schema, we need to know exactly which schema we're working with
without needing to fully validate.

That's a long-winded way of saying that all JSONModel objects now include a
"jsonmodel_type" property that is the name of the model they're expecting to
validate against (like "dog" or "dessert").  All of our existing use-cases will
work out the appropriate type automatically and slot it in, so existing APIs
won't change.  Situations like the above will require the hash
to manually specify a value for "jsonmodel_type" to resolve the ambiguity.