Added some paranoia-based security checks.
When calling .to_jsonmodel, force the caller to specify which repo_id they're
querying.  Support extra options to explicitly say "I don't care which
repository" and "This object exists outside of any repository".

Shouldn't make any semantic difference at the moment, but just adding this to
reduce the chances of mistakes being made in the future.