Refactored to remove the need to pass the currently selected :repo_id around.
Now the initial controller request will set the 'context' for the request, and
all code that needs to know the current repository scope can look to that
context.

In addition, all models now register themselves as either running in a
repository scope or a global scope, so we're able to automatically restrict
them to repositories.