Start each database transaction closer to the point it will be used.
Previously we would establish the database transaction in the backend's
middleware.  This was convenient, but presents a problem: the transaction block
might get retried if the database rolls back our transaction.  The second time
the block was run, the user's request would be re-parsed, and this would fail.

Since we now have all controllers going through our Endpoints, it seems to make
sense to move the database transaction inside the body of the controller--so
the transaction starts as soon as the user request has successfully been
parsed.  This has the nice side effect of making it possible to conditionally
use transactions based on the request type.  For example, maybe we don't bother
for certain GET requests.