Increase the number of retries in the Selenium tests before failing.
Previously the tests would retry for 10 seconds before failing (20
sleeps of 500ms).  When I changed the sleep time to 100ms I only
increased the number of retries to 40, so the time before failing
dropped to 4 seconds.  Increased the number of sleeps to 100 to return
the time before failure to 10 seconds.