Refactored the form templates to allow a form to specify a source of required fields, default values, etc.
Use this facility to bind forms to JSONModel schemas.  This lets us
automatically indicate whether a field is required based on the
schema, pull in values for dropdowns from schema enum definitions, and
things like that.  Also allowed me to get rid of some bits in the
templates that were interrogating the JSON schemas directly.