EAD export: instances for archival_objects
This commit might beprobably sloppier than it should be. However,
for some reason the `ead_containers` method used in the top-level
<did> was not working for archival_objects. There's a slight change
in the tests that add some additional instances to the
archival_objects that get created.

TODO: Add more tests, esp. for DAOs that get exported from records
linked to archival_objects.