Okidokes.  So the jquery.Multifile plugin didn't work in IE.  Soooo... I rewrote our own mutlifile uploader in the simplest way I could.. by allowing the addition of multiple file inputs.
IE8 also doesn't play nice with ajaxed file uploads and the AjaxForm plugin magically reverts to the trusted iframe mechanism for uploading.  This requires some extra response formatting in our controller. Namely, when we detect an iframePOST, we wrap the result in a <textarea>.  The javascript then parses this <textarea> for either the json or html content.