Check if the session is active when submitting a record form.  If not, display the login form... and upon authentication resubmit the form
Refactor the login javascript so it can be used by both the navbar login form and the record form when sessions are lost

Add new modal template to allow for quick ajax modals to render any template partial it is assigned.

Upon reauthenticating, pass back the new CSRF token so we can poke it in the new form and continue.