Add backend endpoint for loction deletion. If location has links then raise an exception.
Add delete to the location frontend toolbar, controller and routes.

If an error is displayed in the linker form (when creating a new record), focus the form on this error.  It was easy to miss.  Also show the error messages for a location on the form (rather than the top level new/edit templates).

Refactor the JSONModelI18nMixin to allow for a single message to be I18n'ed on demand