Use matching bundler version for plugins
If the version of bundler that gets installed in a plugins gems
directory does not match the version shipped with ArchivesSpace
then ArchivesSpace fails to start.

Explicitly setting the bundler version in initialize-plugin.sh to
match the version in the archivesspace gems directory resolves
the issue.

This will also to need to be done for initialize-plugin.bat