Implement support for repository selection stickiness
The current behavior of selecting the first created repository
(available to a user) each time a user logs in is frustrating
and unintuitive.

This is particularly a problem for sites with multiple
repositories where the "primary" repository is not listed first
in the repository selection dropdown menu.

Use a cookie that includes the session username to track the
currently selected repository so that the selection persists
between logins.

By including the username in the cookie maintaining the selection
for different users from the same browser will be handled correctly.