Don't display interface links if proxy address is localhost
The default is localhost and there is no requirement or instruction
to change this value. When ArchivesSpace is being accessed in a non
localhost context the default behavior is unreasonable as the links
will not work. Therefore only display links between interfaces if
the proxy urls are not localhost (sites can choose to override this
however they like).