Update Tomcat setup to decouple solr / indexer
Tomcat setup wasn't brought in line with recent updates to
enable external Solr instance. This makes Tomcat setup equivalent
to standard install with the indexer using a dedicated port.

Note: a future improvement would be to better handle external Solr
with Tomcat. The embedded Solr war file will not be transferred
but the Solr entry in server.xml is present. This produces an
error on startup in the logs but does not appear to otherwise
impact the system.