Fix resource record public display error with breadcrumbs
If Finding Aid Data has status completed but no title value then
the record will fail to display in the public ui. Therefore check
that there is a value or use the resource record title.

This does expose another issue in the archival object context, but
I'll report that in Pivotal Tracker.