Remove regexp match on XML source.
I have an EAD record that seems to go into a spin on the regexp check
when imported.  I didn't dig into why that was (the stack trace showed
it stuck deep within JRuby's regex engine) because it seemed easier just
to get rid of the regular expression.

I think this implementation should have the same effect, and it roughly
halved the time taken to import hornstein.xml in my tests.