Session expired during edit poll redirect
Trigger a redirect to the welcome index action when the user
session has expired during polling. This suppresses the
potentially massive amounts of AccessDeniedException spam
that comes from an expired session in edit mode (which we
see a lot).

If the user had no changes they will be redirected. If the user
had unsaved changes the confirmation navigation alert is
activated. This gives the user a chance to "stay" and attempt
to save their changes after reauthenticating. The alert blocks
polling so it also prevents spam AccessDeniedException messages.