Changes to avoid potential deadlocks when importing multiple files
Steven Majewski did some analysis on some troubles he'd had with EAD
imports and found that importing multiple files would cause an InnoDB
lock timeout that would cause the import job to be marked as failed when
it had actually succeeded.

Lock timeouts aside, this seemed like a bug, because the job status
should really be set to 'completed' in the same transaction that commits
the imported records.  I changed the job runner interface to allow jobs
to call a `success!` method to mark the job as completed within the
current transaction.  I've modified the existing job runners to call
this method (although it's not strictly required).

In testing that, I hit the same InnoDB timeout issue when using multiple
files.  The job watchdog thread (which just updates the timestamp on the
job to tell the rest of the world that it's still running) was suddenly
finding that the job row was locked, and this would cause the timeout to
occur.

In the end, I tracked this down to the code that writes the job's
created URIs to the `job_created_record` table.  It turns out that
inserting a row into `job_created_record` is enough to take a row-level
lock on the corresponding entry in the `job` table--just because there's
a foreign key relationship between the two.  So that was surprising!

To work around this, I now store the list of created URIs in memory for
the duration of the import job run, then when the import finishes, shut
down the watchdog thread, log all created URIs in one hit, and commit
the transaction.  This avoids the import thread competing with the
watchdog for locks on that table.

If storing URIs in memory ends up being no good, I guess another
approach would be to have the watchdog thread write its timestamps to a
separate `running_job` table that wouldn't be at risk of being locked.