On record update, touch the mtimes of all related records
Bobbi Fox noticed an inconsistency in the search indexes as a part of
her PUI work.  She is working with an archival object record that is
linked to a digital object via an instance record.  To access parts of
the digital object for display, she relies on the fact that the archival
object document in Solr contains the resolved version of that digital
object (inlining the digital object in the JSON record for the
instance).

However, when the digital object record was updated, this didn't trigger
the linked archival object to be reindexed.  As a result, the resolved
digital object (embedded in the archival object's indexed document) was
now stale.

Although there is code in the Relationships mixin that's intended to
handle this case, it wasn't aware of nested records.  As a result, the
update to the digital object touches the system_mtime of the instance
nested record, but not the system_mtime of the archival object that
contains that instance.

This change makes this whole process nested record aware.  When a record
is updated, we traverse its relationships to find the records that link
to it, then further traverse the parent/child relationships of any
nested records to get back to a top-level record.