Security updates
Summary:

Reduces the amount of system information available via the UI. Believe
this was added to help support remote installations, but it's including
too much (and I can't see that it's adding much value). As it's not
uncommon to have api keys and such in the environment it's possible to
inadvertently leak sensitive information.

Restricts access to system information so that only the global admin
account has access to it.

Makes it so only the global admin can edit the global admin and other
system accounts (otherwise any 'admin' user can do this by popping in
the path i.e. /users/2/edit).

Prevents any user from becoming the global admin. That account must be
logged into directly, making it the most privileged account in the
system (as it should be).

Prevents group assignment for 'admin' users. They already have all the
privileges so groups are unnecessary, but this prevents a side effect
where an 'admin' user (including the global admin) can lose admin group
permissions if they are assigned to a group by another admin level user.

These changes are for the frontend only, no backend enforcements added.