AR-1413 Don't inadvertently hide instances
When clicking the "Click to load records" button on a resource record
containing many notes and some instances, the instances section would be
collapsed, never to be seen again.

The code used to initialize the notes section wasn't being properly
scoped, so it was applying to the entire form.  Modified this to only
apply to the notes in question.