AR-1413 Second half of fix for "too many" notes
The root cause of the problems was a function like this:

     function initNoteForms($container) {
       ...
       $('.find-something', $container);
       ... more stuff
     }

being called like this:

     initNoteForms(function () {
       ... callback stuff...
     })

The intention was for the callback to fire after the note forms were
initialized, but in fact the callback was firing first.

The problem was that $container should have been a JQuery
object (matching some DOM element), but instead it was a callback
function.  That function got passed as the second argument to
JQuery.find, which attempts to wrap it in a JQuery function (in case you
passed a naked DOM element, I guess).  Since this:

  $(function () { whatever })

is a synonym for document.onready, the function was immediately
executed, and returned `undefined`.  And since a second argument of
`undefined` to jQuery.find means "search entire document", it happily
applied the selector overly broadly without complaining.

So we ended up with the selector seeming to work, and the callback
running backwards.  Yay, JavaScript!

The thing was, the code was actually mostly working with the callback
being run at the wrong time--and was actually relying on running first.
So, I pulled the "should run first" code out and ran it explicitly, then
initialised the notes form as required, then finally fired a callback to
emit the right events.