Rollback last misguided change.  Fix the real bug.
When the user grabs the scrollbar and drops it, we calculate the
waypoint that will be closest to their new scroll position.  For
example, if there are 100 waypoints and they drop the scrollbar to the
80% mark, we'll start looking for waypoints to populate starting from the
80th waypoint.

The populate works backwards from the starting index, then forwards,
each time looking to see if it's hit a waypoint that's close enough to
the viewport to be populated.  If we hit a waypoint that isn't visible,
there's no point continuing the search in that direction: if waypoint 75
isn't visible, waypoints 0..74 won't be either (since they're further up
in the page).

The bug came about because in the backwards scan I would check 80, 79,
78, ... and then, in the forward scan, 80, 81, 82, ....  If our
somewhat-arbitrarily-picked 80th waypoint was NOT visible, we would
immediately terminate searches in both directions, even if one of 81 or
79 might have been visible.

So, we should always check at least one less and one more than the
starting index when looking for waypoints to populate.  Otherwise we
occasionally fail to populate anything.