Include the Relationships mixin in all ASModel classes
(Story time follows, sorry)

If two records are linked by a relationship, a change in one record
causes the system_mtimes of *both* records to be updated.  This makes
indexing work the way you want: if a resource record is linked to a
creator agent record, you want to be able to find the resource by
searching for the creator's name.  If the agent record is updated to
change that name, you want to reindex the resource record to reflect
the new name.

In ArchivesSpace 2.0 we extended this functionality to include nested
records.  So now if a record has a nested record and that nested
record has a relationship with a second record, an update to that
second record will propagate back across the relationship link, then
up the nested hierarchy to the original record.  Butterflies flap
their wings, records get reindexed, etc..

In making this change, we inadvertently assumed that all ASModel
models would include the Relationship mixin, since that houses the
`update_toplevel_mtimes` method that makes all this work.  In
ArchivesSpace core, that assumption is true, but in plugins that might
not be.

So let's make it true!  There's no real cost to including the
Relationships mixin even if you're not using it, and it saves us
having to deal with the exceptional cases.  I've modified ASModel to
include Relationships automatically, and have removed it from the
places that we explicitly included it.