NOJIRA Expand "implied publication" to match the original intention
The idea behind implied publication is to control which linked records
make it into the public interface.  Specifically, we only want to show
subjects, agents and containers that are actually linked from a
published record.

The original implied publication code got some of the way there, but:

  * It didn't handle top containers (which didn't exist when it was
    first written).

  * It would miss records that were suppressed but still marked as
    published.

  * It wouldn't take record hierarchy into account.

The record hierarchy bit is the main challenge.  Say a subject links
to an archival object within a published resource record.  The subject
is therefore "published by implication" and shows up in the public
interface.  If that top-level resource was then unpublished (or
suppressed), the subject should now be marked as unpublished, even
though the archival object it links to hasn't changed.

That is, a linked record's implied publication status can be affected
by any of its ancestors in the tree, and we need to look at those when
doing the calculations.