Add frontend assets to folder in aspace base and add jetty filter to route
This adds some steps to take all frontend precompiled CSS assets and move them into
a folder ( "assets/frontent" ). It also adds a jetty directive to route
requests in /assets to that folder. If an asset is not in that folder ( like is
an asset that cannot be precompiled or whatever ), it then is passed to the
rails asset pipeline.

Why? The goals:

* Try and keep asset requests off Rails/Sprokets and reduce the load
* Try and add another layer that helps with weird URL changes people do in
  their various setups
* Give people an option if they want a proxy to stand in front and serve up the
  assets