Set Enumeration default value to null when value deleted
It is possible to delete / merge a controlled value that was set
as a default (particularly where the value is used by multiple
enumerations) and this crashes things hard.

So, leverage the database to ensure default value is set to null
if the value is deleted (which is a good thing regardless).