Consolidate uses of prefix configuration settings
We'd spawned two sets of settings that were hard to tell apart:

  * public_prefix
  * public_proxy_prefix
  * frontend_prefix
  * frontend_proxy_prefix

The documentation tells people to use the `*_proxy_prefix` variants,
but the code uses the `*_prefix` variants in several places, resulting
in inconsistent handling of prefixes.

We've now unified these settings so that `public_prefix` and
`frontend_prefix` are still supported, but act as aliases for their
`*_proxy_prefix` counterparts.