Don't allow a person to be merged into another agent record if they're linked to an assessment
This actually relaxes an existing restriction.  Previously we wouldn't
allow one record to be merged into another unless both supported the
same set of relationships.  For assessments, this was a problem:
AgentPerson supports an `assessments` relationship (meaning they can
be linked to an assessment record as reviewer), but the other agent
types don't.  But we still want agent merges to work!

So now we only refuse the merge if one of the victim records has data
in a relationship that would be lost in the merge.  If the victim
model supports a relationship but the specific record wasn't using it,
then no harm done.

We still need to make sure we don't merge a record that was linked to
an assessment, but the delete handling already take care of that for
us.  You can't delete an agent that was acting as a reviewer for an
assessment, and when the transfer process tries to delete the record
it'll hit that restriction.  So all is well.