Switch research value to be a regular rating
Introduces the concept of a readonly rating, whose value can't be
edited in the assessment edit form (instead, its value is derived from
other values when the assessment is saved).  Otherwise, it's the same
as any other rating.

This simplifies the reporting quite a lot, since we don't need a bunch
of special cases to handle reporting on "real" ratings vs reporting on
research value.