Update UPGRADING_2.1.0.md
Ran into this problem on a recent upgrade. This is mentioned in the basic upgrade documents, but this calls it out specificity for this version. It's a bit of a PIA because you don't notice the problem right off the bat, only after the index is populated and people start using the search...