Locking down gems ( i18n, loofah, json ) and removing dup line
There a couple of dependencies that are loosely set to version which can
cause problems. rails-html-sanitizer require loofah gem ~> 2.0, but the
2.1.0 release introduced the crass gem, which might be causing a problem
with a space. Locking loofah to 2.0.3 removes crass.

Also there are known issues with aspace and the i18n gem > v0.7.0.

Also locks down all apps to use the same version of json.

Lastly removes a duplicate line of adding some middleware to rack in the
staff ui ( possible issue ).