Fixes JRuby issue with files being loaded into memory on #copy_stream
There's a JRuby issue ( https://github.com/jruby/jruby/issues/4842 )
that causes IOs to be read into memory is using SSL. This causes big
problems if using Solr or the backend with an SSL socket.
This adds a patch to chunk the IO manually, rather then just reading the
whole thing in.