In Public UI Display a "no records found" error if there's no repos
This displays an error page if there are no repository records found in
the public UI. Currently an error is just thrown.

This also DRYs out the "record not found" behavior in the various
controllers. This is all moved to a rescue_from in the application
controller, which displays a flash and redirects back  ( unless it's the
repository controller, which redirects to the "/" to avoid an redirect
loop )