Remove some dead code and use same advance_query_builder as frontend
This removes some apparently dead code in the controllers/concerns (
debugging and tree_api ). This also follows up on the comment left in
the public's
[advance_query_builder](https://github.com/archivesspace/archivesspace/blob/master/public/app/lib/advanced_query_builder.rb#L1-L4)
to use the advance_query_builder in the common directory. Some slight
modifications where made to the Searchable concern to use that version
of the query builder.