Add Capybara + Selenium test to PUI
This adds Capybara/Selenium tests to the PUI.

The capybara test suite runs with the public:test task.
Tests can be added to the features directory. Included in one to test
for the back button issue that was causing the flash to incorrectly
display that there's a general error.

It is configured to allow for Firefox (default), Chrome, and Headless
Chrome. To use Headless chrome, set an environment variable of
SELENIUM_CHROME. To use regular chrome, set an environment variable of
SELENIUM_HEADY_CHROME.