Add some warning messages about pid files when running Selenium tests
Some java versions ( Oracle java 8? ) seem to skip the ensure block when
the rake task is killed. With selenium, this leaves the test server pid
files in tack, which will then cause problems on subsequent runs. This
just adds a more clear warning message that the pid files are there and
that the tests might not run correctly.