bump jruby & use java jrjackson instead of ruby json for performance bump
Bumps jruby version to 9.2.0.0.

This puts [jrjackson](https://github.com/guyboertje/jrjackson) which
uses the java-based jackson JSON parsers, which gives a small
performance improvement.

Also normalizes the use of the ASUtils.json_parse over the JSON.parse.
Also remove the json gem in the Gemfile, since we're now using the
version that ships with jruby.