Standardize multiselect key handling
On a Mac multiselect is performed by holding down meta; on everything
else it's control.  A multiselect isn't supposed to cancel if you
misclick while holding down the key, but on Mac it does because we
were only looking at the control key in that case.

Pull the key checking into a helper method to make sure we always
check both possible keys.