Add new `current_record` helper to return the currently displayed record object
Added to support the new `as_history` plugin, but more generally
useful.

It's often the case in plugins that a partial needs to work in the
context of multiple record types.  As each record type uses a
different instance variable name to hold the current record, plugins
have historically had to resort to hacks like probing to find non-nil
instance variables to determine which record type is in effect.

This patch adds a new `current_record` helper which just returns the
record appropriate for the current context.