Give defined relationship classes names.
Previously these have been anonymous clasees which can make them
tricky to identify when debugging. This change gives each relationship
class a name in the `Relationships` module. For example:

```
  Relationships::ResourceSubject
```

Where the first part `Resource` is the name of the model on which
the relationship is defined, and the second part `Subject` is the
name of the defined relationship.