move code from Sequel Dirty plugin into ASModel to fix conficts
Unfortunately, simply enabling the plugin as per the Sequel docs breaks
saving agent relationships with a date, due to the after_save hook the
plugin re-defines.

Getting around this for now by copy-pasting code from the plugin into
ASModel directly, so the conflicting hook can be omitted/worked into the
hooks as defined in ASModel.