Improve performance of frontend forms
Reworked the error handling to avoid the need for two-passes
across the templates.

When a validation error comes back for a path like
"/dates_of_existence/1/label", we want to show the right field
label in the error message presented to the user.  That means
somehow mapping a path back to an I18n key like "date.label".

Previously, we did this by building up a map of paths to I18n keys
in the ASpaceFormHelper.  As templates render templates render
templates, we build up a FormContext object that knows the
template currently being rendered and the full path back to the
root of the record.  Since it also knows the I18n keys for the
field labels, it was natural to stash this information for later
use.

However, a problem with the approach is that validation error
messages are emitted at the top of the form--before any of the
form elements are produced.  This means that the map we need
hasn't been populated at the point we actually need it.  Past us
punted on this problem by making the whole process two-pass: one
pass to populate the path to I18n map, then another to actually
produce the form.

That worked fine for a while, but the forms have become more
complex over time, and the cost of rendering them twice has
increased.  So (finally, getting to the point), this commit
changes the behavior: emit the form in one pass, then use JS to
pull back the field label and display it when we need it.  I'm not
claiming it's beautiful, but it's definitely faster!