Add support for plugins defining dependencies
Allows us to load them in the correct order and warn if any required
or recommended dependencies are missing.

Ordering is specified in the plugin `config.yml` like this:

     depends_on_plugins:
       - another_plugin
       - yet_another_plugin
     recommends_plugins:
       - optional_plugin
       - another_optional_plugin

This ensures that your plugin is loaded *after* all of its
dependencies in cases where that matters.  System startup is the
obvious example here, but it also applies to database migrations too.