Show an error message when the RDE hits a Sequel validation error
The RDE will show errors related to JSONModel, but it also needs to show
things that pop up when the records are put into the DB. For example,
DO components require unique identifiers. If the RDE hits an issue when
trying to save, this will slightly modify the error message and just
display the message in the RDE error flash.