Improve enumeration value caching
The series_system plugin makes heavy use of value_for_id calls across
multiple enumerations, and this case was almost always missing the
BackendEnumSource cache.  Change the method to give it a chance of
hitting the cache.

Also increased the cache capacity to compensate for larger numbers of
values.  It's all hit so heavily it's probably not worth tossing out
entries if we can avoid it.