Checkpoint the periodic indexer at the end of each record type
Previously we would defer commit and checkpointing indexer state until
the whole index round finished (for all record types).

When doing a full reindex, this can mean you need to restart from
scratch if anything goes wrong, and that can make you sad.  It's also
not that useful deferring the commit because Solr will autocommit
periodically anyway, so it's not really transactional anyway.

So now we'll commit and record our indexer state as we finish the new
records for each record type.