Make the indexer more robust against partial failures
Sometimes life doesn't go your way, but that doesn't mean you should
throw a tantrum.

The periodic indexer is built around the notion of "index rounds".
Each round figures out which records were changed since the previous
round, grabs those records, maps them into Solr documents and feeds
them to Solr.  As each round finishes successfully, we record a
checkpoint to note that we're officially up-to-date as of this moment.

In its original design, if any of those steps failed, the indexer made
no attempt to recover, and would just detect the error and abort the
index round without writing a checkpoint.  The hope was that errors
would be transient, and that retrying during the next index round had
some chance of succeeding.  It's a blunt instrument, but at least
upholds a "no record left behind" policy--making it pretty obvious
when something is amiss and ensuring that everything gets properly
indexed once the underlying issue is resolved.

An unfortunate side-effect of the "bail out and try again later"
approach is that indexing stops until the underlying problem is
fixed.  This commit attempts to strike a balance between keeping
indexing running and the "no record left behind" policy.  Here's how
it works now:

  * Indexer worker threads will attempt to handle errors while
    fetching and indexing records.  If a batch fails to fetch/index, a
    worker thread will revert to trying one record at a time, indexing
    as much of the batch as possible.

  * When a worker detects an error, it will note the error but
    continue handling subsequent batches of records.  At the end of
    the index round, each worker returns a status which will either
    be:

      - I had nothing to do

      - I successfully indexed everything

      - I encountered one or more errors

  * The periodic indexer will skip its checkpoint at the end of an
    index round if any worker reported an error.

The upshot of all of this is that an index round will continuously
retry from the same checkpoint timestamp until everything indexes
correctly, but everything that *can* be indexed *will* be indexed.

Note that this does make monitoring logs and/or indexer_state
timestamps more important: if a record is causing indexing to fail,
each index round is going to an increasing amount of work as it
re-indexes everything from the last checkpoint time to present.