Fix a bug where the transfer dropdown is closed when the linker modal is closed.
This causes confusion for the user. They click transfer on the tree toolbar for an
archival object. The dropdown shows with the resource linker. If the user then
selects browse from the linker menu, when they close the linker modal, either by
linking to a resource or canceling, the transfer dropdown also closes. The user
thinks 'what just happened?'. If they click the transfer dropdown button again
they will see the linker, populated or not correctly, but by now they've lost
all faith that things will go well for them.

This fixes the problem. It ensures that the transfer dropdown remains open after
the linker modal is closed, so the user can see the result of their linker
adventure, and can go ahead and complete the transfer, or not, as they desire.